requirements.txt
# ffmpeg needed in system
telethon
requests
python-dotenv
fastapi
uvicorn[standard]